Move code to google/cloud/oauth2
